requirements.txt
pyaudio==0.2.12
numpy==1.23.4
scipy==1.9.3
pyyaml==6.0
tqdm==4.64.1
librosa==0.9.2
